fix test for reconciling, add reverse translate logic for controller
